# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Etscript API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/check:
    get:
      operationId: getAdminCheck
      tags: [admin]
      summary: Check whether the current user has admin access
      responses:
        "200":
          description: User is an admin
          content:
            application/json:
              schema:
                type: object
                required: [isAdmin]
                properties:
                  isAdmin:
                    type: boolean
        "403":
          description: Not an admin
...
  /admin/plans:
    get:
      operationId: getAdminPlans
      tags: [admin]
      summary: List all plans (including inactive)
      responses:
        "200":
          description: All plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createAdminPlan
      tags: [admin]
      summary: Create a new plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanInput"
      responses:
        "201":
          description: Created plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /admin/plans/{planId}:
    patch:
      operationId: updateAdminPlan
      tags: [admin]
      summary: Update a plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanInput"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deleteAdminPlan
      tags: [admin]
      summary: Delete a plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
...
components:
  schemas:
...
    Plan:
      type: object
      required: [id, slug, name, description, priceKobo, billingPeriod, isActive, isFeatured, maxManuscripts, maxStorageMb, features, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        priceKobo:
          type: integer
        billingPeriod:
          type: string
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        maxManuscripts:
          type: integer
        maxStorageMb:
          type: integer
        features:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    CreatePlanInput:
      type: object
      required: [slug, name, priceKobo, billingPeriod, maxManuscripts, maxStorageMb]
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        priceKobo:
          type: integer
        billingPeriod:
          type: string
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        maxManuscripts:
          type: integer
        maxStorageMb:
          type: integer
        features:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
    UpdatePlanInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        priceKobo:
          type: integer
        billingPeriod:
          type: string
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        maxManuscripts:
          type: integer
        maxStorageMb:
          type: integer
        features:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
...