# lib/api-zod/src/generated/types/updatePlanInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Etscript API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePlanInput {
  name?: string;
  description?: string;
  priceKobo?: number;
  billingPeriod?: string;
  isActive?: boolean;
  isFeatured?: boolean;
  maxManuscripts?: number;
  maxStorageMb?: number;
  features?: string[];
  sortOrder?: number;
}